feat: Add mapping to track funded users
This commit adds a mapping to track funded users in the PamojaFi contract. The mapping, named `s_fundedUsers`, is a private mapping that associates user addresses with a boolean value indicating whether they have been funded. This mapping is used in the `fund` function to mark users as funded when they receive a deposit. The addition of this mapping enhances the functionality of the contract by providing a way to track funded users.

Co-authored-by: ybtuti <samuelmsuu@gmail.com>
Co-authored-by: Dancan254 <dancanian25@gmail.com>