feat: Add OracleLib library for checking Chainlink Oracle data
This commit adds the OracleLib library to the contracts/src/libraries directory. The OracleLib library provides functions for checking the freshness of data from a Chainlink Oracle. It includes functions to check the time since the last update, retrieve the latest round data, and perform a stale check on the data. The library also defines a constant TIMEOUT value for determining the maximum allowed time since the last update.