feat: Update PamojaFi contract with OracleLib for checking Chainlink Oracle data
This commit updates the PamojaFi contract in the contracts/src/PamojaFi.sol file. It adds the OracleLib library from the contracts/src/libraries/OracleLib.sol file to provide functions for checking the freshness of data from a Chainlink Oracle. The functions include checking the time since the last update, retrieving the latest round data, and performing a stale check on the data. This update enhances the functionality of the contract by ensuring the accuracy and reliability of the price feed data.