feat: Set & Verify wallet address using Coinbase identity dynamically in CreateProposalForm
This commit modifies the CreateProposalForm component in the frontend to set the wallet address dynamically. Previously, the wallet address was hardcoded, but now it is fetched from the connected Ethereum accounts and stored in the state using the setWalletAddress function. This change improves the functionality of the form by ensuring that the correct wallet address is used when submitting the form.

Refactor package.json to update dependencies
Refactor package.json to add new dependencies
Refactor projectHeaderImageLink validation to improve URL validation
Merge pull request #15 from ybtuti/main